use wgpu4k release (#11)
* Fix macOS architecture mapping in build script

Corrected the mapping of macOS architectures to their respective targets. X86_64 now maps to macosX64, and AARCH64 maps to macosArm64. This ensures proper configuration for macOS builds based on architecture.

* Update library versions in gradle configuration

Upgraded wgpu4k to stable 0.0.3 version and coroutines to 1.10.1. This ensures compatibility and utilizes the latest improvements and fixes from the updated dependencies. Configuration